Split fs.go to smaller files
